+++ client-course-schedulizer/src/components/Tabs/Tabs.scss
@@ -1,1 +1,7 @@
+.import-button {
+  margin-right: 1em !important;
+}
 
+.import-button * {
+  color: white !important;
+}
