+++ client-course-schedulizer/src/components/reuseables/AddSectionButton/AddSectionButton.tsx
@@ -1,19 +1,20 @@
-import { Button, IconButton, Popover } from "@material-ui/core";
+import { Button, ButtonProps, IconButton, Popover } from "@material-ui/core";
 import { Add } from "@material-ui/icons";
 import { bindPopover, bindTrigger, usePopupState } from "material-ui-popup-state/hooks";
 import React from "react";
 import { AddSectionPopover } from "../AddSectionPopover";
 import "./AddSectionButton.scss";
 
-interface AddSectionButton {
+interface AddSectionButton extends ButtonProps {
   isIcon?: boolean;
 }
 
-export const AddSectionButton = ({ isIcon }: AddSectionButton) => {
+export const AddSectionButton = (props: AddSectionButton) => {
   const popupState = usePopupState({
     popupId: "addSection",
     variant: "popover",
   });
+  const { isIcon, ...buttonProps } = props;
 
   return (
     <>
@@ -22,7 +23,7 @@ export const AddSectionButton = ({ isIcon }: AddSectionButton) => {
           <Add />
         </IconButton>
       ) : (
-        <Button color="primary" variant="contained" {...bindTrigger(popupState)}>
+        <Button color="secondary" variant="contained" {...buttonProps} {...bindTrigger(popupState)}>
           Add Section
         </Button>
       )}
